Add `jekyll preview` command to render pages on demand
Adds a `preview` subcommand to Jekyll via its commands API (which uses
Mercenary). `jekyll preview` is similar to `jekyll serve`, but instead
of rendering the whole site by default every time the command is
called, it doesn't render anything until a request is made, and then
it only renders the requested page.

This is accomplished in two parts: first, extending Jekyll's existing
Serve command (and it's internal 'Servlet') to pass the current URL
path to Jekyll's Regenerator (part of the existing incremental
regeneration feature), and then extending Regenerator to tell Jekyll
that it should only render and write the file matching the URL.

This approach has the advantage of relying largely on Jekyll's already
tested features. Notably, the stages of Jekyll's build process prior
to render still function on the whole site, so pages that rely on data
loaded from other files (like an index of blog posts) should still
render correctly.

At the moment this is a very alpha release; it's only really been
tested with a default `jekyll new` site. It also calls the build
command from start to finish for every request, though it's probable
that it could at least cache the Site instance rather than create a
new one for every request, and possible that it could perform the
earlier build stages (like read) only once.